% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_marquee.R
\docType{class}
\name{GeomMarquee}
\alias{GeomMarquee}
\title{Geom class for geom_marquee}
\usage{
GeomMarquee
}
\description{
This is the underlying Geom class that powers \code{\link[=geom_marquee]{geom_marquee()}} layers.
}
\keyword{internal}
