% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guide_marquee.R
\docType{class}
\name{GuideMarquee}
\alias{GuideMarquee}
\title{Guide class for guide_marquee}
\usage{
GuideMarquee
}
\description{
This is sthe underlying Guide class that powers \link{guide_marquee} legends.
}
\keyword{internal}
