% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marquee-package.R
\docType{package}
\name{marquee-package}
\alias{marquee}
\alias{marquee-package}
\title{marquee: Markdown Parser and Renderer for R Graphics}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides the mean to parse and render markdown text with grid along with facilities to define the styling of the text.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://marquee.r-lib.org}
  \item \url{https://github.com/r-lib/marquee}
  \item Report bugs at \url{https://github.com/r-lib/marquee/issues}
}

}
\author{
\strong{Maintainer}: Thomas Lin Pedersen \email{thomas.pedersen@posit.co} (\href{https://orcid.org/0000-0002-5147-4711}{ORCID})

Authors:
\itemize{
  \item Martin Mitáš (Author of MD4C)
}

Other contributors:
\itemize{
  \item Posit Software, PBC (03wc8by49) [copyright holder, funder]
}

}
\keyword{internal}
