% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_marquee.R
\name{GeomMarquee}
\alias{GeomMarquee}
\title{Geom class for geom_marquee}
\usage{
GeomMarquee
}
\description{
This environment contains the Geom needed for \code{\link[=geom_marquee]{geom_marquee()}} in the \code{geom}
variable. It is set up like this to avoid ggplot2 to be a hard dependency of
marquee
}
\keyword{internal}
