% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_set.R
\name{style_set}
\alias{style_set}
\alias{modify_style}
\alias{remove_style}
\title{Create or modify a style set that describes a full markdown text}
\usage{
style_set(...)

modify_style(style_set, tag, ...)

remove_style(style_set, tag)
}
\arguments{
\item{...}{Named arguments providing a style for the specific tags. For
\code{modify_style()} a number of style options to change. If the first argument
is a marquee style it will overwrite the tag and subsequent arguments are
ignored}

\item{style_set}{A style set to modify}

\item{tag}{The name of a tag to modify or remove}
}
\value{
A style set object
}
\description{
A style set contains information on how to style the various tags in a
markdown text. While it is not necessary to provide a style for all tags (it
will just inherit the parent if missing), it is required to provide a
complete style for the body tag so an option is avialable through inheritance
for all tags and all style options. It can often be easier to derive a new
style set from an existing one rather than building one from scratch.
}
\examples{
style_set(body = body_style(), p = style(indent = em(2))) |>
  modify_style("p", size = 16) |>
  modify_style("str", style(weight = "bold")) |>
  remove_style("p")

}
